modified deploy script to check and ensure the npm run build exits 0

+++ deploy.sh
@@ -9,10 +9,12 @@ npm install
 
 npm run build
 
-cd build
-
-aws s3 rm s3://${AWS_ENV}.build.magickpics.com --recursive
-echo "Doing a sync to the bucket"
-aws s3 sync . s3://${AWS_ENV}.build.magickpics.com
-
-echo "Done!"
+if [[ $? ne 0 ]]; then
+  exit 1
+else
+  cd build
+  aws s3 rm s3://${AWS_ENV}.build.magickpics.com --recursive
+  echo "Doing a sync to the bucket"
+  aws s3 sync . s3://${AWS_ENV}.build.magickpics.com
+  echo "Done!"
+fi
